fix: private route
